api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["Nathan Segers <nathan.segers@howest.be>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.88.0"
sqlalchemy = "^1.4.45"
pymysql = "^1.0.2"
python-dotenv = "^0.21.0"
strawberry-graphql = {version = "0.144.3", extras = ["debug-server"]}
pymongo = {extras = ["srv"], version = "^4.3.3"}


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
